Move ogn client (gateway) to a separate file.
The documentation[1] says:
> In the simplest case, __init__.py can just be an empty file,
> but it can also execute initialization code for the package or
> set the __all__ variable, described later.

A class like 'ognGateway' can't be considered 'initialization code'.

[1]: https://docs.python.org/3/tutorial/modules.html#packages
